+++ src/Domain/Common/AuditEntity.cs
@@ -4,7 +4,7 @@ namespace Grocery.Domain.Entities
 {
     public class AuditEntity
     {
-        public int CreatedById { get; set; }
+        public long CreatedById { get; set; }
         public User CreatedBy { get; set; }
         public DateTimeOffset CreatedAt { get; set; }
         public DateTimeOffset UpdatedAt { get; set; }
