+++ src/Domain/Entities/Brand.cs
@@ -2,9 +2,9 @@ using Grocery.Domain.Common;
 
 namespace Grocery.Domain.Entities
 {
-    public class Brand : IEntity<int>
+    public class Brand : IEntity<long>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public bool IsPublished { get; set; }
