+++ src/Domain/Entities/Role.cs
@@ -1,12 +1,11 @@
-using System;
 using System.Collections.Generic;
 using Grocery.Domain.Common;
 
 namespace Grocery.Domain.Entities
 {
-    public class Role : IEntity<int>
+    public class Role : IEntity<long>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public List<UserRole> UserRoles { get; set; }
