+++ src/Domain/Entities/UserAddress.cs
@@ -3,11 +3,11 @@ using Grocery.Domain.Enums;
 
 namespace Grocery.Domain.Entities
 {
-    public class UserAddress
+    public class UserAddress 
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
-        public int AddressId { get; set; }
+        public long AddressId { get; set; }
         public Address Address { get; set; }
         public AddressType AddressType { get; set; }
         public DateTimeOffset? LastUsedOn { get; set; }
